Big UI Changes plus xp implemented
Added a custom font, using tmpro. Changed most text to use it. Enemies drop xp that goes to the player, leveling up the weapon at milestones. Weapon sprite and level plus %xp shown in hud.

Items can now be picked up on the map, and items purchased from the shop are removed from the pool. If no items are left in the pool a default will be chosen when restocking. Once an item is purchased, a sold out img is displayed. KNOWN BUG: this will display even if the item is not purchased due to lack of gold.
Shop also displays current player gold (cached).

There is a github project board tracking all intended future changes!